# apiserver/model_sensor.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Xovis app API
 *
 * API to access and configure the Xovis app
 *
 * API version: 1.0.0
 */

package apiserver

type Sensor struct {
	Id int32 `json:"id,omitempty"`

	ConfigurationId int32 `json:"configuration_id,omitempty"`

	Username string `json:"username,omitempty"`

	Password string `json:"password,omitempty"`

	Hostname string `json:"hostname,omitempty"`

	Port int32 `json:"port,omitempty"`

	DiscoveryMode string `json:"discovery_mode,omitempty"`

	L3FirstIp *string `json:"l3_first_ip,omitempty"`

	L3Count *int32 `json:"l3_count,omitempty"`
}

// AssertSensorRequired checks if the required fields are not zero-ed
func AssertSensorRequired(obj Sensor) error {
	return nil
}

// AssertSensorConstraints checks if the values respects the defined constraints
func AssertSensorConstraints(obj Sensor) error {
	return nil
}