# apiserver/model_sensor_create_update.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Xovis app API
 *
 * API to access and configure the Xovis app
 *
 * API version: 1.0.0
 */

package apiserver

type SensorCreateUpdate struct {
	Id int32 `json:"id,omitempty"`

	ConfigurationId int32 `json:"configuration_id"`

	Username string `json:"username"`

	Password string `json:"password"`

	Hostname string `json:"hostname"`

	Port int32 `json:"port"`

	DiscoveryMode string `json:"discovery_mode"`

	L3FirstIp *string `json:"l3_first_ip,omitempty"`

	L3Count *int32 `json:"l3_count,omitempty"`
}

// AssertSensorCreateUpdateRequired checks if the required fields are not zero-ed
func AssertSensorCreateUpdateRequired(obj SensorCreateUpdate) error {
	elements := map[string]interface{}{
		"configuration_id": obj.ConfigurationId,
		"username":         obj.Username,
		"password":         obj.Password,
		"hostname":         obj.Hostname,
		"port":             obj.Port,
		"discovery_mode":   obj.DiscoveryMode,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertSensorCreateUpdateConstraints checks if the values respects the defined constraints
func AssertSensorCreateUpdateConstraints(obj SensorCreateUpdate) error {
	return nil
}